--- src/Layout/Header/Header.css
@@ -1,119 +1,0 @@
-.header {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 30px 20px 28px 20px;
-}
-.header__logo-section {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.header__logo-section-logo {
-  width: 70px;
-  height: 70px;
-}
-@media screen and (max-width: 820px) {
-  .header__logo-section-logo {
-    width: 68px;
-    height: 68px;
-  }
-}
-.header__logo-section__btnChangeTheme {
-  padding-left: 1.1rem;
-  cursor: pointer;
-}
-@media screen and (max-width: 820px) {
-  .header__logo-section__btnChangeTheme {
-    height: 28px;
-  }
-}
-.header__navigation {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.7rem;
-}
-@media screen and (max-width: 720px) {
-  .header__navigation {
-    display: none;
-  }
-}
-.header__navigation__ul {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  gap: 25px;
-  list-style: none;
-  padding: 0;
-}
-.header__navigation__ul-nav-link {
-  position: relative;
-  text-decoration: none;
-  font-size: 111.11%;
-  line-height: 130%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  transition: color 0.3s ease-in-out;
-}
-.header__navigation__ul-nav-link::after {
-  content: "";
-  position: absolute;
-  left: 0;
-  bottom: -4px;
-  width: 100%;
-  height: 2px;
-  background-color: #393;
-  transform: scaleX(0);
-  transform-origin: bottom right;
-  transition: transform 0.3s ease-in-out;
-}
-.header__navigation__ul-nav-link:hover {
-  color: #393;
-}
-.header__navigation__ul-nav-link:hover::after {
-  transform: scaleX(1);
-  transform-origin: bottom left;
-}
-@media screen and (max-width: 820px) {
-  .header__navigation__ul-nav-link {
-    font-size: 88.89%;
-  }
-}
-.header__action__ul {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 266.67%;
-  gap: 1rem;
-}
-.header__action__ul .header__action__ul-item.icon {
-  font-size: 48px;
-}
-@media screen and (max-width: 1000px) {
-  .header__action__ul .header__action__ul-item.icon {
-    font-size: 44px;
-  }
-}
-@media screen and (max-width: 480px) {
-  .header__action__ul .header__action__ul-item.icon {
-    font-size: 32px;
-  }
-}
-.header__action__ul .header__action__ul-item.menu {
-  cursor: pointer;
-  border: none;
-  display: none;
-}
-.header__action__ul .header__action__ul-item.menu:hover {
-  opacity: 0.7;
-}
-@media screen and (max-width: 720px) {
-  .header__action__ul .header__action__ul-item.menu {
-    display: block;
-  }
-}/*# sourceMappingURL=Header.css.map */
