--- src/components/Banner/Banner.css
@@ -1,92 +1,0 @@
-.banner {
-  position: relative;
-  height: 600px;
-  width: 100%;
-  background-image: url("./banner.png");
-  background-size: cover;
-  background-position: bottom;
-  background-repeat: no-repeat;
-}
-.banner::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.4);
-}
-@media screen and (max-width: 1000px) {
-  .banner {
-    height: 476px;
-  }
-}
-@media screen and (max-width: 768px) {
-  .banner {
-    height: 634px;
-  }
-}
-@media screen and (max-width: 480px) {
-  .banner {
-    height: 600px;
-  }
-}
-.banner__content {
-  position: relative;
-  z-index: 1;
-  margin: auto 40px;
-  height: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: left;
-  top: 12%;
-}
-.banner__content__header {
-  font-size: 96px;
-  color: #fff;
-  line-height: 110%;
-  font-weight: 700;
-}
-@media screen and (max-width: 1000px) {
-  .banner__content__header {
-    font-size: 90px;
-  }
-}
-@media screen and (max-width: 910px) {
-  .banner__content__header {
-    font-size: 68px;
-  }
-}
-@media screen and (max-width: 480px) {
-  .banner__content__header {
-    font-size: 60px;
-  }
-}
-@media screen and (max-width: 360px) {
-  .banner__content__header {
-    font-size: 52px;
-  }
-}
-.banner__content__checkOutBtn {
-  padding: 16px 56px;
-  background: #393;
-  color: #fff;
-  font-family: Montserrat;
-  max-width: -moz-fit-content;
-  max-width: fit-content;
-  font-size: 20px;
-  line-height: 130%;
-  font-weight: 600;
-  border: none;
-  border-radius: 6px;
-  margin-top: 40px;
-  cursor: pointer;
-}
-@media screen and (max-width: 480px) {
-  .banner__content__checkOutBtn {
-    font-size: 18px;
-  }
-}
-.banner__content__checkOutBtn:hover {
-  background-color: rgb(42, 131, 42);
-}/*# sourceMappingURL=Banner.css.map */
