+++ src/components/Banner/Banner.scss
@@ -3,7 +3,7 @@
   position: relative;
   height: 600px;
   width: 100%;
-  background-image: url("./banner.png");
+  background-image: url("../../media/images/banner.png");
   background-size: cover;
   background-position: bottom;
   background-repeat: no-repeat;
