--- src/components/DiscountButton/DiscountButton.css
@@ -1,16 +1,0 @@
-.discount-button {
-  color: white;
-  border: none;
-  padding: 6px 16px;
-  border-radius: 6px;
-  background-color: #393;
-  max-width: -moz-fit-content;
-  max-width: fit-content;
-  font-size: 89%;
-  line-height: 126%;
-  cursor: pointer;
-  transition: color 0.4s ease;
-}
-.discount-button:hover {
-  background-color: rgb(42, 131, 42);
-}/*# sourceMappingURL=DiscountButton.css.map */
