--- src/components/ProductCard/style.module.css
@@ -1,72 +1,0 @@
-.productCard {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-  align-items: center;
-}
-
-.cardBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.productCard a {
-  color: black;
-  text-decoration: none;
-}
-
-.titleP {
-  font-size: 12px;
-}
-
-.productImg {
-  width: 100%;
-  height: 250px;
-  -o-object-fit: contain;
-     object-fit: contain;
-  border-radius: 15px;
-}
-
-.productCard button {
-  width: 225px;
-  height: 50px;
-  border: 2px solid #339933;
-  background-color: #F1FFF1;
-  color: #339933;
-  font-size: 15px;
-  font-weight: 600;
-  border-radius: 20px;
-  cursor: pointer;
-  opacity: 0;
-  transition: 0.4s ease-out;
-}
-
-.productCard:hover button {
-  display: block;
-  opacity: 1;
-  transform: translateY(-150px);
-}
-
-.priceBlock {
-  width: 200px;
-  display: flex;
-  flex-direction: row;
-  align-items: flex-end;
-  justify-content: space-between;
-}
-
-.priceP {
-  font-weight: 600;
-  font-size: 25px;
-}
-
-.oldPriceP {
-  text-decoration: line-through;
-  color: #8B8B8B;
-}
-
-.percentDiscountP {
-  color: #FF32A1;
-  font-weight: 600;
-}/*# sourceMappingURL=style.module.css.map */
