+++ src/index.js
@@ -1,74 +1,70 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import App from './App'
 import './style/app.scss'
-import { Provider } from 'react-redux';
-import  {store} from './store/store';
-import { persistor } from "./store/store";
-import { PersistGate } from "redux-persist/integration/react";
-import MainPage from "./pages/MainPage";
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-import CategoriesPage from './pages/CategoriesPage';
-import ProductsPage from './pages/ProductsPage/ProductsPage';
-import SingleProductPage from "./pages/SingleProductPage/SingleProductPage"
-import SalesPage from './pages/SalesPage/SalesPage';
-import ErrorPage from './pages/ErrorPage';
-import SingleCategoryPage from './pages/SingleCategoryPage';
-
-
+import { Provider } from 'react-redux'
+import { store } from './store/store'
+import { persistor } from './store/store'
+import { PersistGate } from 'redux-persist/integration/react'
+import MainPage from './pages/MainPage'
+import { RouterProvider, createBrowserRouter } from 'react-router-dom'
+import CategoriesPage from './pages/CategoriesPage'
+import ProductsPage from './pages/ProductsPage'
+import SingleProductPage from './components/SingleProductPage/SingleProductPage'
+import SalesPage from './pages/SalesPage'
+import SingleCategoryPage from './pages/SingleCategoryPage'
+import SingleProduct from './pages/SingleProduct'
 
 const router = createBrowserRouter([
   {
-    path: "/", // домашняя страница
+    path: '/', // домашняя страница
     element: <App />,
     children: [
       {
-        path: "/",
+        path: '/',
         element: <MainPage />,
       },
       {
-        path: "/categories",
+        path: '/categories',
         element: <CategoriesPage />,
       },
       {
-        path: "/categories/:id",
+        path: '/categories/:id',
 
         element: <SingleCategoryPage />,
-
       },
       {
-        path: "/products",
+        path: '/products',
         element: <ProductsPage />,
       },
       {
-        path: "/products/:id",
+        path: '/products/:id',
 
-        element: <SingleProductPage />,
+        element: <SingleProduct />,
 
         // element: <SingleProductPage />,
-
       },
       {
-        path: "/sales",
+        path: '/sales',
         element: <SalesPage />,
       },
       {
-        path: "/cart",
+        path: '/cart',
         // element: <CartPage />,
       },
       {
-        path: "/favorites",
+        path: '/favorites',
         // element: <FavoritesPage />,
       },
       {
-        path: "*",
-        element: <ErrorPage />,
+        path: '*',
+        // element: <ErrorPage />,
       },
     ],
   },
-]);
+])
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
   <Provider store={store}>
     <PersistGate persistor={persistor}>
@@ -76,5 +72,5 @@ root.render(
         <App />
       </RouterProvider>
     </PersistGate>
-  </Provider>
-);
+  </Provider>,
+)
