--- src/pages/ErrorPage.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function ErrorPage() {
-  return <div>Page Not Found</div>;
-}
