--- src/pages/SingleProductPage.jsx
@@ -1,14 +1,0 @@
-// import React from "react";
-
-// import BreadCrumbs from "./../components/BreadCrumbs/BreadCrumbs";
-
-// const SingleProductPage = () => {
-//   return (
-//     <section className="container">
-//       <BreadCrumbs />
-//       <div>SingleProductPage</div>
-//     </section>
-//   );
-// };
-
-// export default SingleProductPage;
