--- src/pages/SingleProductPage/style.module.css
@@ -1,140 +1,0 @@
-.card {
-  width: 1110px;
-  margin: 60px auto;
-  display: grid;
-  grid-template-columns: 600px 450px;
-}
-
-.header {
-  display: flex;
-  padding-bottom: 20px;
-  font-weight: 600;
-  font-size: 20px;
-  color: #3a3a3a;
-}
-
-.rightBlock {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 50px;
-}
-
-.leftBlock {
-  display: flex;
-  flex-direction: column;
-}
-
-.leftBlock img {
-  width: 500px;
-  -o-object-fit: contain;
-     object-fit: contain;
-  cursor: pointer;
-}
-
-.priceBlock {
-  width: 200px;
-  display: flex;
-  flex-direction: row;
-  align-items: flex-end;
-  justify-content: space-between;
-  gap: 100px;
-}
-
-.priceP {
-  font-weight: 600;
-  font-size: 50px;
-}
-
-.oldPriceP {
-  text-decoration: line-through;
-  color: #8b8b8b;
-}
-
-.percentDiscountP {
-  color: #ff32a1;
-  font-weight: 600;
-  font-size: 20px;
-}
-
-.descriptionBlock {
-  border-top: 2px dashed rgba(192, 192, 192, 0.7529411765);
-  padding-top: 20px;
-}
-
-.descriptionBlock #header {
-  font-weight: 600;
-  padding-bottom: 20px;
-}
-
-.buttonsContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 10px;
-}
-
-.countButtonContainer {
-  box-sizing: border-box;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  border: none;
-  border-top: 1px solid gray;
-  border-bottom: 1px solid gray;
-  height: 100%;
-  width: 40%;
-  border-radius: 10px;
-  cursor: pointer;
-}
-
-.countValue {
-  padding: 0 10%;
-  color: #8b8b8b;
-}
-
-.countButton {
-  border-top: none;
-  border-bottom: none;
-  border-left: 1px solid gray;
-  border-right: 1px solid gray;
-  color: #8b8b8b;
-  height: 58px;
-  width: 58px;
-  border-radius: 10px;
-  cursor: pointer;
-  font-size: 2rem;
-}
-
-.addToCartButton {
-  flex: 2;
-  width: 232px;
-  height: 58px;
-  border: none;
-  background-color: #339933;
-  color: white;
-  font-size: 20px;
-  font-weight: 600;
-  border-radius: 10px;
-  cursor: pointer;
-}
-
-.modalBackdrop {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-  cursor: pointer;
-}
-
-.modalImage {
-  max-width: 80%;
-  max-height: 80%;
-  cursor: pointer;
-}/*# sourceMappingURL=style.module.css.map */
